+++ src/app/core/modules/custom-router-dom/AuthRoute.tsx
@@ -4,6 +4,7 @@ import { useNavigate } from 'react-router-dom';
 import { AppRoutes } from '@app/core/constants/app-routes';
 import { authStorage } from '@app/core/services/auth-storage.service';
 
+// check logged in - if logged in, not navigate to auth page
 export const AuthRoute = ({
   component: Wrapper,
 }: {
