+++ src/app/core/modules/custom-router-dom/PrivateRoute.tsx
@@ -4,6 +4,7 @@ import { Navigate } from 'react-router-dom';
 import { authStorage } from '@app/core/services/auth-storage.service';
 import { AppRoutes } from '@app/core/constants/app-routes';
 
+// checked logged in, if not logged in, navigate to auth
 export const PrivateRoute = ({
   component: Wrapped,
 }: {
