+++ src/app/shared/components/layouts/Header.tsx
@@ -21,12 +21,13 @@ export const Header = () => {
   const dropdownRef = useRef<HTMLDivElement>(null);
   const lastScrollY = useRef(0);
   const [hidden, setHidden] = useState(false);
-  const navigate = useNavigate();
 
+  // check is authen
   const dispatch = useAppDispatch();
   const user = useAppSelector((state) => state.auth.user);
   const isAuthenticated = !!user;
 
+  // check gender to set default icon avatar
   const author =
     user?.gender === 'female'
       ? femaleIcon
@@ -34,6 +35,7 @@ export const Header = () => {
       ? maleIcon
       : otherIcon;
 
+  // scroll header
   useEffect(() => {
     const handleScroll = () => {
       const currentScrollY = window.scrollY;
@@ -48,6 +50,7 @@ export const Header = () => {
     return () => window.removeEventListener('scroll', handleScroll);
   }, []);
 
+  // click outsite to set close dropdown
   useEffect(() => {
     const handleClickOutside = (event: MouseEvent) => {
       if (
@@ -61,6 +64,7 @@ export const Header = () => {
     return () => document.removeEventListener('mousedown', handleClickOutside);
   }, []);
 
+  // logout
   const onLogoutConfirm = () => {
     dispatch(logoutThunk());
     toast.success('Logout successfully');
