+++ src/app/store/hook/useAppSelector.ts
@@ -1,4 +1,5 @@
 import { TypedUseSelectorHook, useSelector } from 'react-redux';
+
 import type { RootState } from '../index';
 
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
