+++ src/app/store/post/reducers/postReducer.ts
@@ -1,4 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit';
+
 import {
   fetchPostsThunk,
   createPostThunk,
